* tag 'v2.12.1':
  sync: pass --bare option when doing git clone of superproject.
  sync: superproject performance changes.
  manifest: set revisionId as revision attribute it it is not being set in ToXml.
  Update _CheckForImmutableRevision to use git rev-list
  sync: Added --filter=blob:none for git clone of superproject.
  sync: Do a git pull with --use-superproject if superproject tree already exists.
  sync: Added --filter=blob:none (and no-depth) wduring git clone of superproject.
  git_command.py: Handle unicode decode error
  sync: Disable info about disabling pruning when quiet

+++ git_superproject.py
@@ -25,9 +25,12 @@ Examples:
 import os
 import sys
 
-from error import GitError
+from error import BUG_REPORT_URL, GitError
 from git_command import GitCommand
 import platform_utils
+
+_SUPERPROJECT_GIT_NAME = 'superproject.git'
+_SUPERPROJECT_MANIFEST_NAME = 'superproject_override.xml'
 
 
 class Superproject(object):
@@ -47,6 +50,10 @@ class Superproject(object):
     self._repodir = os.path.abspath(repodir)
     self._superproject_dir = superproject_dir
     self._superproject_path = os.path.join(self._repodir, superproject_dir)
+    self._manifest_path = os.path.join(self._superproject_path,
+                                       _SUPERPROJECT_MANIFEST_NAME)
+    self._work_git = os.path.join(self._superproject_path,
+                                  _SUPERPROJECT_GIT_NAME)
 
   @property
   def project_shas(self):
@@ -58,12 +65,14 @@ class Superproject(object):
 
     Args:
       url: superproject's url to be passed to git clone.
-      branch: the branchname to be passed as argument to git clone.
+      branch: The branchname to be passed as argument to git clone.
 
     Returns:
       True if 'git clone <url> <branch>' is successful, or False.
     """
-    cmd = ['clone', url, '--depth', '1']
+    if not os.path.exists(self._superproject_path):
+      os.mkdir(self._superproject_path)
+    cmd = ['clone', url, '--filter', 'blob:none', '--bare']
     if branch:
       cmd += ['--branch', branch]
     p = GitCommand(None,
@@ -80,6 +89,29 @@ class Superproject(object):
       return False
     return True
 
+  def _Fetch(self):
+    """Do a 'git fetch' to to fetch the latest content.
+
+    Returns:
+      True if 'git fetch' is successful, or False.
+    """
+    if not os.path.exists(self._work_git):
+      print('git fetch missing drectory: %s' % self._work_git,
+            file=sys.stderr)
+      return False
+    cmd = ['fetch', 'origin', '+refs/heads/*:refs/heads/*', '--prune']
+    p = GitCommand(None,
+                   cmd,
+                   cwd=self._work_git,
+                   capture_stdout=True,
+                   capture_stderr=True)
+    retval = p.Wait()
+    if retval:
+      print('repo: error: git fetch call failed with return code: %r, stderr: %r' %
+            (retval, p.stderr), file=sys.stderr)
+      return False
+    return True
+
   def _LsTree(self):
     """Returns the data from 'git ls-tree -r HEAD'.
 
@@ -88,14 +120,15 @@ class Superproject(object):
     Returns:
       data: data returned from 'git ls-tree -r HEAD' instead of None.
     """
-    git_dir = os.path.join(self._superproject_path, 'superproject')
-    if not os.path.exists(git_dir):
-      raise GitError('git ls-tree. Missing drectory: %s' % git_dir)
+    if not os.path.exists(self._work_git):
+      print('git ls-tree missing drectory: %s' % self._work_git,
+            file=sys.stderr)
+      return None
     data = None
     cmd = ['ls-tree', '-z', '-r', 'HEAD']
     p = GitCommand(None,
                    cmd,
-                   cwd=git_dir,
+                   cwd=self._work_git,
                    capture_stdout=True,
                    capture_stderr=True)
     retval = p.Wait()
@@ -108,25 +141,29 @@ class Superproject(object):
           retval, p.stderr), file=sys.stderr)
     return data
 
-  def GetAllProjectsSHAs(self, url, branch=None):
+  def _GetAllProjectsSHAs(self, url, branch=None):
     """Get SHAs for all projects from superproject and save them in _project_shas.
 
     Args:
-      url: superproject's url to be passed to git clone.
-      branch: the branchname to be passed as argument to git clone.
+      url: superproject's url to be passed to git clone or fetch.
+      branch: The branchname to be passed as argument to git clone or fetch.
 
     Returns:
       A dictionary with the projects/SHAs instead of None.
     """
     if not url:
       raise ValueError('url argument is not supplied.')
-    if os.path.exists(self._superproject_path):
-      platform_utils.rmtree(self._superproject_path)
-    os.mkdir(self._superproject_path)
 
-    # TODO(rtenneti): we shouldn't be cloning the repo from scratch every time.
-    if not self._Clone(url, branch):
-      raise GitError('git clone failed for url: %s' % url)
+    do_clone = True
+    if os.path.exists(self._superproject_path):
+      if not self._Fetch():
+        # If fetch fails due to a corrupted git directory, then do a git clone.
+        platform_utils.rmtree(self._superproject_path)
+      else:
+        do_clone = False
+    if do_clone:
+      if not self._Clone(url, branch):
+        raise GitError('git clone failed for url: %s' % url)
 
     data = self._LsTree()
     if not data:
@@ -147,3 +184,67 @@ class Superproject(object):
 
     self._project_shas = shas
     return shas
+
+  def _WriteManfiestFile(self, manifest):
+    """Writes manifest to a file.
+
+    Args:
+      manifest: A Manifest object that is to be written to a file.
+
+    Returns:
+      manifest_path: Path name of the file into which manifest is written instead of None.
+    """
+    if not os.path.exists(self._superproject_path):
+      print('error: missing superproject directory %s' %
+            self._superproject_path,
+            file=sys.stderr)
+      return None
+    manifest_str = manifest.ToXml().toxml()
+    manifest_path = self._manifest_path
+    try:
+      with open(manifest_path, 'w', encoding='utf-8') as fp:
+        fp.write(manifest_str)
+    except IOError as e:
+      print('error: cannot write manifest to %s:\n%s'
+            % (manifest_path, e),
+            file=sys.stderr)
+      return None
+    return manifest_path
+
+  def UpdateProjectsRevisionId(self, manifest, projects, url, branch=None):
+    """Update revisionId of every project in projects with the SHA.
+
+    Args:
+      manifest: A Manifest object that is to be written to a file.
+      projects: List of projects whose revisionId needs to be updated.
+      url: superproject's url to be passed to git clone or fetch.
+      branch: The branchname to be passed as argument to git clone or fetch.
+
+    Returns:
+      manifest_path: Path name of the overriding manfiest file instead of None.
+    """
+    try:
+      shas = self._GetAllProjectsSHAs(url=url, branch=branch)
+    except Exception as e:
+      print('error: Cannot get project SHAs for %s: %s: %s' %
+            (url, type(e).__name__, str(e)),
+            file=sys.stderr)
+      return None
+
+    projects_missing_shas = []
+    for project in projects:
+      path = project.relpath
+      if not path:
+        continue
+      sha = shas.get(path)
+      if sha:
+        project.SetRevisionId(sha)
+      else:
+        projects_missing_shas.append(path)
+    if projects_missing_shas:
+      print('error: please file a bug using %s to report missing shas for: %s' %
+            (BUG_REPORT_URL, projects_missing_shas), file=sys.stderr)
+      return None
+
+    manifest_path = self._WriteManfiestFile(manifest)
+    return manifest_path
